refactored RootReducer
