+++ package.json
@@ -19,7 +19,7 @@
     "react-dom": "^16.12.0",
     "react-redux": "^7.1.3",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.2.0",
+    "react-scripts": "^3.3.0",
     "react-stripe-checkout": "^2.6.3",
     "react-stripe-elements": "^6.0.1",
     "redux": "^4.0.4",
